requirements.txt
streamlit
pandas
matplotlib
google-genai
python-dotenv
